+++ pkg/sass-parser/lib/src/expression/index.ts
@@ -7,6 +7,7 @@ import type {
   BinaryOperationExpression,
   BinaryOperationExpressionProps,
 } from './binary-operation';
+import {BooleanExpressionProps} from './boolean';
 import type {StringExpression, StringExpressionProps} from './string';
 
 /**
@@ -14,14 +15,17 @@ import type {StringExpression, StringExpressionProps} from './string';
  *
  * @category Expression
  */
-export type AnyExpression = BinaryOperationExpression | StringExpression;
+export type AnyExpression =
+  | BinaryOperationExpression
+  | StringExpression
+  | BooleanExpressionProps;
 
 /**
  * Sass expression types.
  *
  * @category Expression
  */
-export type ExpressionType = 'binary-operation' | 'string';
+export type ExpressionType = 'binary-operation' | 'string' | 'boolean';
 
 /**
  * The union type of all properties that can be used to construct Sass
@@ -31,7 +35,8 @@ export type ExpressionType = 'binary-operation' | 'string';
  */
 export type ExpressionProps =
   | BinaryOperationExpressionProps
-  | StringExpressionProps;
+  | StringExpressionProps
+  | BooleanExpressionProps;
 
 /**
  * The superclass of Sass expression nodes.