+++ pkg/sass-parser/CHANGELOG.md
@@ -22,6 +22,7 @@
 ## 0.2.5
 
 * Add support for parsing the `@supports` rule.
+* Add `BooleanExpression` and `NumberExpression`.
 
 ## 0.2.4